# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Documentation",
        "license": {
            "name": "API"
        },
        "version": "1.0.0"
    },
    "basePath": "/public/api",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/chemicals/addtobookmark": {
            "post": {
                "tags": [
                    "Chemicals"
                ],
                "summary": "Add to user bookmark  \u0434\u043e\u0431\u0430\u0432\u0438\u0442\u044c \u0432 \u0437\u0430\u043a\u043b\u0430\u0434\u043a\u0438",
                "description": "Add to user bookmark",
                "operationId": "Add to user bookmark",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "chemical_id",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
...
        "/decorator/addtobookmark": {
            "post": {
                "tags": [
                    "Decorator"
                ],
                "summary": "Add to user bookmark",
                "description": "Add to user bookmark",
                "operationId": "Add to user bookmark",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "decorator_id",
                        "in": "query",
                        "description": "decorator ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
...
        "/decorator/deletefrombookmark": {
            "post": {
                "tags": [
                    "Decorator"
                ],
                "summary": "delete from user handbook bookmark",
                "description": "delete from user handbook bookmark",
                "operationId": "delete from user handbook bookmark",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "handbook_id",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
...
        "/disease/addtobookmark": {
            "post": {
                "tags": [
                    "Disease"
                ],
                "summary": "Add to user bookmark",
                "description": "Add to user bookmark",
                "operationId": "Add to user bookmark",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "disease_id",
                        "in": "query",
                        "description": "disease ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
...
        "/event/addtobookmark": {
            "post": {
                "tags": [
                    "Event"
                ],
                "summary": "Add to user bookmark",
                "description": "Add to user bookmark",
                "operationId": "Add to user bookmark",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "query",
                        "description": "event ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
...
        "/handbook/index": {
            "post": {
                "tags": [
                    "Handbook"
                ],
                "summary": "Get handbooks",
                "description": "Get handbooks",
                "operationId": "Get handbooks",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "section_id",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "25 56  6-7 - attribute values",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/handbook/show/{id}": {
            "post": {
                "tags": [
                    "Handbook"
                ],
                "summary": "Get handbook",
                "description": "Get handbook",
                "operationId": "Get handbook",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "handbook ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/handbook/addvideolink": {
            "post": {
                "tags": [
                    "Handbook"
                ],
                "summary": "Work in POSTMAN",
                "description": "Add video link",
                "operationId": "Work in POSTMAN",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "handbook_id",
                        "in": "formData",
                        "description": "handbook ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "description": "link's title",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "link",
                        "in": "formData",
                        "description": "link",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "photo",
                        "in": "formData",
                        "description": "\u041d\u0430\u0444\u0438\u0433 \u043d\u0435 \u043d\u0430\u0434\u043e, \u043d\u043e \u0431\u0435\u0437 \u043d\u0435\u0433\u043e \u043d\u0435 \u0440\u0430\u0431\u043e\u0442\u0430\u0442\u0435",
                        "required": false,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/handbook/addtobookmark": {
            "post": {
                "tags": [
                    "Handbook"
                ],
                "summary": "Add to user bookmark",
                "description": "Add to user bookmark",
                "operationId": "Add to user bookmark",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "handbook_id",
                        "in": "query",
                        "description": "handbook ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/handbook/checkisbookmark": {
            "post": {
                "tags": [
                    "Handbook"
                ],
                "summary": "check is handbook in users bookmark",
                "description": "check is handbook in users bookmark",
                "operationId": "check is handbook in users bookmark",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "handbook_id",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/handbook/deletefrombookmark": {
            "post": {
                "tags": [
                    "Handbook"
                ],
                "summary": "delete from user handbook bookmark",
                "description": "delete from user handbook bookmark",
                "operationId": "delete from user handbook bookmark",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "handbook_id",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/handbook/showfilter": {
            "post": {
                "tags": [
                    "Handbook"
                ],
                "summary": "show handbook filter",
                "description": "show handbook filter",
                "operationId": "show handbook filter",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "section_id",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/handbook/addarticle": {
            "post": {
                "tags": [
                    "Handbook"
                ],
                "summary": "Add article",
                "description": "Add particle",
                "operationId": "Add article",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "article",
                        "in": "formData",
                        "description": "doc, docs, pdf or other",
                        "required": true,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
...
        "/pest/addtobookmark": {
            "post": {
                "tags": [
                    "Pests"
                ],
                "summary": "Add to user bookmark",
                "description": "Add to user bookmark",
                "operationId": "Add to user bookmark",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "pest_id",
                        "in": "query",
                        "description": "pest ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
...
        "/question/addtobookmark": {
            "post": {
                "tags": [
                    "QuestionAnswer"
                ],
                "summary": "Add to user bookmark",
                "description": "Add to user bookmark",
                "operationId": "Add to user bookmark",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "question_id",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
...
        "/sort/addtobookmark": {
            "post": {
                "tags": [
                    "Sorts"
                ],
                "summary": "Add to user bookmark  \u0434\u043e\u0431\u0430\u0432\u0438\u0442\u044c \u0432 \u0437\u0430\u043a\u043b\u0430\u0434\u043a\u0438",
                "description": "Add to user bookmark",
                "operationId": "Add to user bookmark",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "sort_id",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "access_token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}